I was looking into improving the performance of my application so I set up a scenario with `BenchmarkDotNet` which, using that particular schema, evaluated 1000 json documents. These changes reduce the time from 465 ms to 220 ms.

Here are the results from running `JsonSchema.Benchmark` on my machine.

Before:
```
|          Method |  n |       Mean |   Error |  StdDev |       Gen0 |       Gen1 | Allocated |
|---------------- |--- |-----------:|--------:|--------:|-----------:|-----------:|----------:|
| RunSuite_Legacy |  1 |   523.1 ms | 5.43 ms | 4.82 ms | 10000.0000 |  5000.0000 |  127.6 MB |
| RunSuite_Legacy | 10 | 1,249.4 ms | 6.08 ms | 5.08 ms | 63000.0000 | 16000.0000 | 754.56 MB |
```

After:
```
|          Method |  n |     Mean |   Error |  StdDev |       Gen0 |      Gen1 | Allocated |
|---------------- |--- |---------:|--------:|--------:|-----------:|----------:|----------:|
| RunSuite_Legacy |  1 | 481.6 ms | 3.95 ms | 3.69 ms |  8000.0000 | 4000.0000 | 101.88 MB |
| RunSuite_Legacy | 10 | 793.2 ms | 5.69 ms | 5.33 ms | 42000.0000 | 5000.0000 | 506.87 MB |
```