@gregsdennis 2023-12-26
Thanks for looking into this.  These look like good optimizations overall.